efclub_django/requirements.txt
Django==4.0.4
gunicorn==20.1.0
loguru==0.6.0
psycopg2==2.9.3
python-dotenv==0.20.0

djangorestframework==3.13.1
Markdown==3.3.7
django-filter==22.1
